move submit.js
