+++ src/login/Submit.js
@@ -1,5 +1,5 @@
-import { Button } from "./Button";
-import { login, loginCheck} from '../login/LoginCheck';
+import { Button } from "../common/Button";
+import { login, loginCheck} from './LoginCheck';
 
 export function SubmitButton(item) {
   return Button({
